3rd commit, added deletion functionality
Added deletion functionality so that it works properly. Fixed a bug where the class that holds the list would still there, leading to a build up of empty white spaces. Also fixed a bug that deleted only the most recent item, no matter what item you pressed the delete button on. The core functionality is now done, and just the styling & the checkmark text styling is left to do.